pyproject.toml
[tool.poetry]
name = "assignment1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.38.0"
pandas = "^2.2.2"
numpy = "^2.1.1"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"
requests = "^2.32.3"
openai = "^1.46.0"
sqlalchemy = "^2.0.35"
psycopg2-binary = "^2.9.9"
pytest = "^8.3.3"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
black = "^24.8.0"
isort = "^5.13.2"
mypy = "^1.11.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
